Moved a few PUT routes from the reservationsController to the resRoomsController. Updated the API docs accordingly.
